+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Fixed an issue jumping to definitions in syntactically invalid
 code. Previously we assumed that we could always macroexpand code,
 which isn't true on unfinished code.
 
+Fixed a crash when starting from a symbol that was at the very start
+or end of a buffer.
+
 # v1.0
 
 Initial release.